[320] Add CourseSearchService
CourseSearchService is now only responsible for fetching records and not
ordering them.

Rework the specs to use real data rather than mocking scopes. The
mocking approach has proved really difficult to maintain when we needed
to update the scopes and add includes.

Only add the outer scope when provider name is present. This should
hopefully give some performance benefits. The only reason the outer
scope is needed is because the original query contains duplicates and
`.distinct` errors due to the `Course#accredited_body_order scope which
adds an anonymouse order field to the query that can't be grouped by. We
should look to rework that but I've left it for now.