[320] Add CourseOrderingService
This service accepts a course scope and filter and sort strings and
orders them accordingly.

It was extracted from the CourseSearchService as part of work to fix a
pagination bug. Turns out the bug was in the canonical scopes on
`Course` but this class is complex and will be easier to maintain on its
own.

Also add some more descriptive naming to help make the intentions
clearer in the code.

Re-do the tests using real data. The mocking approach has grown difficult to
maintain and is of questionable value now the class has grown more
complex.